Added getConfig( kClass: KClass< out OpMode > )
This will be how the delegation for configurables is done (see Logging).

For this change to work, I had to rewrite profileFor to search for the
TeleOp and Autonomous annotations, extract the name value from them, then
call another method which will create the OpModeConfig if it doesn't
already exist (I realized there was a bug where this would cause no new
OpModeConfigs to be created).