Uses AnnotatedName when fetching profile
Instead of using annotations and reflections magic in yet another
spot and dealing with errors in more places, Configurations.profileFor
now accepts a KAbstractOpMode object (so, 'this' when it's called)
and uses its AnnotatedOpMode property to get the name of the config
to get the active profile for.

This is a much nicer looking system and delegates the error to the
KAbstractOpMode class, where the actual problem is, instead of the
Configurations object.