Added overhead for RpmLockedMotoer
This will be what allows motors to try to output the given RPM as much as
it can, but this also blocks all updates to any DcMotor properties from
the user-end. That's the deal with this class: You can have a motor
constantly trying to reach a given RPM, but you will not be allowed to
modify anything other than the target output rpm.