Added LoggedVariable
A variable delegated to a LoggedVariable< T > will write a log statement
every time its value is read or written. This is useful to trace the use
of important variables over time and keep it in the log so that the flow
of the program can be determined after its has completed.