Reworked Hardware System
UtilityContainer & FalseHardwareMap are no longer needed, as the OpModes
have been reverted to not having a hardwaremap initialized immediately
when run.

K(Linear)OpMode has lost its hardware-accessing get method, as that has
been delegate to the HardwareDefinition class (and this reinforces using a
reusable HardwareDefinition instead of inheritance or copying and
pasting).

HardwareDefinition now requires an OpMode as a parameter, this is so that
we can access the hardware map again. This won't be much of a change in
syntax, so this should be fine.